llkn: colocate TransmissionKeys with its serialization
Move WriteTo/ReadFrom/BinarySize/MarshalBinary/UnmarshalBinary from
llkn/serialization.go into llkn/keys.go next to the TransmissionKeys
struct, then delete llkn/serialization.go. Same change as the previous
kgplus commit.
